* VM.cs: * Utils.cs: * vmor.cs: * vmnor.cs: * vmdiv.cs: * vmjmp.cs: * vmmov.cs: * vmmul.cs: * vmret.cs: * vmneg.cs: * vmpop.cs: * vmclr.cs: * vmfbd.cs: * vmfbi.cs: * vmxor.cs: * vmadd.cs: * vmand.cs: * vmsub.cs: * vmxnor.cs: * vmcpu.csproj: * vmpush.cs: * vmpeek.cs: * vmcall.cs: * vmnand.cs: * MemoryMap.cs: * test.asm: * Frame.cs: * Program.cs: * vmoperator.cs: * vmcli.csproj: * vmasm.csproj: * Assembler.cs: * packages.config: * Register.cs: * Instruction.cs: * FramebufferForm.cs: * gui.stetic:
* Framebuffer.cs: Add Framebuffer delegates
Add Own FB ram int

* Memory.cs: Erbt nun von MemoryStream

* vmcpu.inc: Add new Prototypes

+++ Komponent/Register.cs
@@ -140,7 +140,7 @@ namespace Vcsos.Komponent
 			case "IP":
 				return ip;
 			case "TIK":
-				return CPU.Ticks;
+				return (int)CPU.Ticks;
 			default:
 				throw new Exception ("Register/Flag with Name: " + name + " not found");
 			}
